[refactor] #3 : deprecated된 코드 수정
