+++ filchat/src/main/java/com/project/filchat/infrastructure/jwt/JwtProvider.java
@@ -9,13 +9,12 @@ import org.springframework.stereotype.Component;
 
 import com.project.filchat.common.exception.ErrorCode;
 import com.project.filchat.common.exception.UnAuthorizedException;
-import com.project.filchat.domain.auth.AuthDto;
 import com.project.filchat.infrastructure.config.properties.JwtProperties;
+import com.project.filchat.interfaces.auth.AuthDto;
 
 import io.jsonwebtoken.ExpiredJwtException;
 import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.security.Keys;
 
 @Component
@@ -30,35 +29,35 @@ public class JwtProvider {
         this.refreshTokenExpirationTime = jwtProperties.refreshTokenExpirationTime();
     }
 
-    public AuthDto.LoginResponse createJwtToken(String userToken, String nickname) {
+    public AuthDto.TokenResponse createJwtToken(String userToken, String nickname) {
         Date now = new Date();
         Date accessTokenExpiration = new Date(now.getTime() + accessTokenExpirationTime);
         Date refreshTokenExpiration = new Date(now.getTime() + refreshTokenExpirationTime);
 
         String accessToken = Jwts.builder()
-            .signWith(secretKey, SignatureAlgorithm.HS256)
-            .setIssuedAt(now)
-            .setExpiration(accessTokenExpiration)
+            .signWith(secretKey, Jwts.SIG.HS256)
+            .issuedAt(now)
+            .expiration(accessTokenExpiration)
             .claim("userToken", userToken)
             .claim("nickname", nickname)
             .compact();
 
         String refreshToken = Jwts.builder()
-            .signWith(secretKey, SignatureAlgorithm.HS256)
-            .setIssuedAt(now)
-            .setExpiration(refreshTokenExpiration)
+            .signWith(secretKey, Jwts.SIG.HS256)
+            .issuedAt(now)
+            .expiration(refreshTokenExpiration)
             .claim("userToken", userToken)
             .claim("nickname", nickname)
             .compact();
-        return new AuthDto.LoginResponse(userToken, accessToken, refreshToken);
+        return new AuthDto.TokenResponse(userToken, accessToken, refreshToken);
     }
 
     public void validateToken(final String token) {
         try {
             Jwts.parser()
-                .setSigningKey(secretKey)
+                .verifyWith(secretKey)
                 .build()
-                .parseClaimsJws(token);
+                .parseSignedClaims(token);
         } catch (ExpiredJwtException e) {
             throw new UnAuthorizedException(ErrorCode.EXPIRED_TOKEN);
         } catch (JwtException e) {
